Minor update - with an important bug fix
Several minor improvements for computing speed. The major improvement is with the SingleParticleAnalyzerV2 and ParticlePairAnalyerV2 which have been redesign to fill two-particle histograms in 2D rather than 3D.
This update also include an important bug fix that was causing memory leaks and growing arrays in the Nucleus class. I was looking in the wrong place for 3 weeks before I finally manage to identify the source of the issue in DefineAs method.